requirements.txt
streamlit 
langchain
langchain_core 
langchain_community 
langchain_groq
python-dotenv 
faiss-cpu
langchain_huggingface 
langchain_text_splitters 
PyPDF2 
docx
torch
openai-whisper
chromadb
sentence-transformers
sounddevice
numpy 
scipy 
PyPDF2 
transformers
huggingface_hub